requirements.txt
pytest==7.4.3
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
typer==0.9.0
rich==13.7.0